add generic queries for up to 8 components
